create a basic webpack + express server
This includes building Webpack and express, and then hosting the files
from Express when Express is running in production. When running
locally, Webpack reverse proxies to the Express server, so its static
assets can be refreshed dynamically.

This is very basic. Babel is included for React support. There is no
Babel transformations for the server yet.